Removes links to Discourse
Not knowing the way the site is built, and having failed to set up local
hosting of it to test how it looks, this is a bit of a "search-and-destroy"
approach to things. If not a PR to merge, let's consider this a reminder PR
for what needs to be done.

Also: might be worth us adding a Discord link all over the place.